Integration of model w/ android
The java file implements many of the functions in the python package. It also includes a spellchecker that i developed for use in the android app. The python files are mainly utilities to make the integration easier or improve understanding rather than actual code for the app, but the java file is meant to be part of the Android source code. Also, these files assume the presence of a 16M-word Arabic dictionary that was too large to upload here but which you can find at this link: https://sourceforge.net/projects/arabic-wordlist/

This is still a work in progress (I was making slow progress because of school, but now that I'm done with digital logic (the hardest class) I'm planning to spend most of my available time on this to get it into production as soon as possible) but hopefully you get the idea of what im trying to do